requirements.txt
ansicolors==1.0.2
getch==1.0
